Refresh v4  from latest master

+++ DynamicData/List/Internal/Sort.cs
@@ -4,28 +4,29 @@ using System.Linq;
 using System.Reactive;
 using System.Reactive.Linq;
 using DynamicData.Annotations;
+using DynamicData.Internal;
 using DynamicData.Kernel;
 
-namespace DynamicData.Internal
+namespace DynamicData.List.Internal
 {
     internal sealed class Sort<T>
     {
         private readonly IObservable<IChangeSet<T>> _source;
-        private  IComparer<T> _comparer;
+        private IComparer<T> _comparer;
         private readonly SortOptions _sortOptions;
         private readonly int _resetThreshold;
         private readonly IObservable<Unit> _resort;
         private readonly IObservable<IComparer<T>> _comparerObservable;
 
 
-        public Sort([NotNull] IObservable<IChangeSet<T>> source, 
+        public Sort([NotNull] IObservable<IChangeSet<T>> source,
             [NotNull] IComparer<T> comparer, SortOptions sortOptions,
-            IObservable<Unit> resort, 
+            IObservable<Unit> resort,
             IObservable<IComparer<T>> comparerObservable,
             int resetThreshold)
         {
             if (source == null) throw new ArgumentNullException(nameof(source));
-            if (comparer == null) throw new ArgumentNullException(nameof(comparer));
+            //if (comparer == null) throw new ArgumentNullException(nameof(comparer));
             _source = source;
             _comparer = comparer;
             _sortOptions = sortOptions;
@@ -41,19 +42,20 @@ namespace DynamicData.Internal
                 var locker = new object();
                 var orginal = new ChangeAwareList<T>();
                 var target = new ChangeAwareList<T>();
-                
+
                 var changed = _source.Synchronize(locker).Select(changes =>
                 {
                     if (_resetThreshold > 1)
                         orginal.Clone(changes);
 
-                    return changes.TotalChanges > _resetThreshold ? Reset(orginal, target) : Process(target, changes);
-                    
+                    return changes.TotalChanges > _resetThreshold && _comparer!=null ? Reset(orginal, target) : Process(target, changes);
                 });
                 var resort = _resort.Synchronize(locker).Select(changes => Reorder(target));
                 var changeComparer = _comparerObservable.Synchronize(locker).Select(comparer => ChangeComparer(target, comparer));
 
-                return changed.Merge(resort).Merge(changeComparer).SubscribeSafe(observer);
+                return changed.Merge(resort).Merge(changeComparer)
+                .Where(changes=>changes.Count!=0)
+                .SubscribeSafe(observer);
             });
         }
 
@@ -72,54 +74,60 @@ namespace DynamicData.Internal
 
         private IChangeSet<T> ProcessImpl(ChangeAwareList<T> target, IChangeSet<T> changes)
         {
+            if (_comparer == null)
+            {
+                target.Clone(changes);
+                return target.CaptureChanges();
+            }
+
             foreach (var change in changes)
             {
                 switch (change.Reason)
                 {
                     case ListChangeReason.Add:
-                    {
-                        var current = change.Item.Current;
-                        Insert(target, current);
-                        break;
-                    }
+                        {
+                            var current = change.Item.Current;
+                            Insert(target, current);
+                            break;
+                        }
                     case ListChangeReason.AddRange:
-                    {
-                        var ordered = change.Range.OrderBy(t => t, _comparer).ToList();
-                        if (target.Count == 0)
                         {
+                            var ordered = change.Range.OrderBy(t => t, _comparer).ToList();
+                            if (target.Count == 0)
+                            {
                                 target.AddRange(ordered);
+                            }
+                            else
+                            {
+                                ordered.ForEach(item => Insert(target, item));
+                            }
+                            break;
                         }
-                        else
-                        {
-                            ordered.ForEach(item => Insert(target, item));
-                        }
-                        break;
-                    }
                     case ListChangeReason.Replace:
-                    {
-                        var current = change.Item.Current;
-                        //TODO: check whether an item should stay in the same position
-                        //i.e. update and move
-                        Remove(target, change.Item.Previous.Value);
-                        Insert(target, current);
-                        break;
-                    }
+                        {
+                            var current = change.Item.Current;
+                            //TODO: check whether an item should stay in the same position
+                            //i.e. update and move
+                            Remove(target, change.Item.Previous.Value);
+                            Insert(target, current);
+                            break;
+                        }
                     case ListChangeReason.Remove:
-                    {
-                        var current = change.Item.Current;
-                        Remove(target, current);
-                        break;
-                    }
+                        {
+                            var current = change.Item.Current;
+                            Remove(target, current);
+                            break;
+                        }
                     case ListChangeReason.RemoveRange:
-                    {
+                        {
                             target.RemoveMany(change.Range);
-                        break;
-                    }
+                            break;
+                        }
                     case ListChangeReason.Clear:
-                    {
+                        {
                             target.Clear();
-                        break;
-                    }
+                            break;
+                        }
                 }
             }
             return target.CaptureChanges();
@@ -127,20 +135,20 @@ namespace DynamicData.Internal
 
         private IChangeSet<T> Reorder(ChangeAwareList<T> target)
         {
-                int index = -1;
-                var sorted = target.OrderBy(t => t, _comparer).ToList();
-                foreach (var item in sorted)
-                {
-                    index++;
+            int index = -1;
+            var sorted = target.OrderBy(t => t, _comparer).ToList();
+            foreach (var item in sorted)
+            {
+                index++;
 
-                    var existing = target[index];
-                   //if item is in the same place, 
-                    if (ReferenceEquals(item, existing)) continue;
+                var existing = target[index];
+                //if item is in the same place, 
+                if (ReferenceEquals(item, existing)) continue;
 
-                    //Cannot use binary search as Resort is implicit of a mutable change
-                    var old = target.IndexOf(item);
-                    target.Move(old, index);
-                }
+                //Cannot use binary search as Resort is implicit of a mutable change
+                var old = target.IndexOf(item);
+                target.Move(old, index);
+            }
 
             return target.CaptureChanges();
         }
